requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Pillow==10.4.0
numpy==1.26.4
h5py==3.11.0
tensorflow==2.17.1
ultralytics>=8.2.0,<9.0.0
openai>=1.0.0
gunicorn==22.0.0
